dev(x86): change project init
