+++ src/arch/x86/x86.c
@@ -1,7 +1,11 @@
 #include "internal/x86.h"
+#include <memory/memory.h>
 
-void x86_init(void)
+void x86_init(uint32_t magic, uint32_t mbi_addr)
 {
 	gdt_init();
 	idt_init();
+	mb2_init(magic, mbi_addr);
+	page_descriptor_init();
+	buddy_init();
 }
